style: #231 스푼 컴포넌트 추가
